django/requirements.txt
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.5.0
djangorestframework-jwt==1.11.0
django-cleanup==5.1.0
python-jose==3.2.0
psycopg2-binary==2.8.6
opencv-python-headless==4.4.0.44
cryptography==3.1.1
tensorflow==2.3.0
Keras==2.4.3
Pillow==7.2.0
